backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
python-multipart==0.0.6
httpx==0.24.1
redis==5.0.1
python-dotenv==1.0.0
pandas==2.1.4
reportlab==4.0.7
supabase==2.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
gunicorn==21.2.0